Refactor and rename SCons/Util/subprocess_context.py to SCons/Util/contexts.py.
Changes:
* Split SubprocessContext into ProcessContext and ShellContext and associated warnings.
* Base classes are no longer abstract and serve as non-windows handlers.
* Method names changed from get_warning_message to update_* and returns warning message.
* Refactored command interpreter prefix for windows generated commands.
